+++ routers/app_v1.py
@@ -20,6 +20,7 @@ limitations under the License.
 from app import utils
 from app import rate_limiter
 
+from routers.v1.math import app_v1_math
 from routers.v1.language import app_v1_language
 
 from fastapi import APIRouter, Request, BackgroundTasks
@@ -27,7 +28,8 @@ from fastapi import APIRouter, Request, BackgroundTasks
 # Initializing the "app_v1" API Router
 app_v1 = APIRouter(prefix="/v1")
 
-# Include the API Routers
+# Including the API Routers
+app_v1.include_router(app_v1_math)
 app_v1.include_router(app_v1_language)
 
 # Route 1: Main (app_v1)